Rearranged some gui module
