+++ internal/gui/widgets/play_bar.go
@@ -1,4 +1,4 @@
-package gui
+package widgets
 
 import (
 	"fyne.io/fyne/v2"
