+++ internal/gui/widgets/thumbnail.go
@@ -1,4 +1,4 @@
-package gui
+package widgets
 
 import (
 	"io"
