+++ internal/gui/window_app/theme.go
@@ -1,4 +1,4 @@
-package gui
+package window_app
 
 import (
 	"image/color"
