+++ internal/i18n/translations/zh_CN/ui.yaml
@@ -27,3 +27,6 @@
 
 - Key: wrapper_adder
   Default: "由{{.Adder}}点歌"
+
+- Key: app_name
+  Default: "VRC跳舞房预加载器"
